Пофикшены ошибки сборки x86

+++ seospider/src/viewport_percent_resize_policy.cpp
@@ -28,7 +28,7 @@ int ViewportPercentResizePolicy::columnSize(int column, const TableView* view) c
 	const int scrollBarWidth = view->verticalScrollBar()->isVisible() ? view->verticalScrollBar()->width() : 0;
 	const int width = view->width() - scrollBarWidth - s_extraWidth;
 
-	DEBUG_ASSERT(m_columnsPercentSize.size() > static_cast<size_t>(column));
+	DEBUG_ASSERT(m_columnsPercentSize.size() > column);
 
 	const int columnWidth = m_columnsPercentSize[column] * width / 100;
